Make generate_ule_validators code more useful
