+++ utils/zerovec/derive/src/ule.rs
@@ -13,10 +13,6 @@ use syn::{
     Fields, Ident, Lit,
 };
 
-fn suffixed_ident(name: &str, suffix: usize, s: Span) -> Ident {
-    Ident::new(&format!("{name}_{suffix}"), s)
-}
-
 pub fn derive_impl(input: &DeriveInput) -> TokenStream2 {
     if !utils::has_valid_repr(&input.attrs, |r| r == "packed" || r == "transparent") {
         return Error::new(
@@ -337,24 +333,10 @@ pub(crate) fn generate_ule_validators<'a>(
     iter: impl Iterator<Item = &'a Field>,
     // (validators, remaining_offset)
 ) -> (TokenStream2, syn::Ident) {
-    let mut prev_offset_ident = Ident::new("ZERO", Span::call_site());
-    let mut validators = quote!(const ZERO: usize = 0);
-
-    for (i, field) in iter.enumerate() {
+    utils::generate_per_field_offsets(iter, |field, prev_offset_ident, size_ident, _| {
         let ty = &field.ty;
-        let new_offset_ident = suffixed_ident("OFFSET", i, field.span());
-        let size_ident = suffixed_ident("SIZE", i, field.span());
-        validators = quote! {
-            #validators;
-            const #size_ident: usize = ::core::mem::size_of::<#ty>();
-            const #new_offset_ident: usize = #prev_offset_ident + #size_ident;
-            <#ty as zerovec::ule::ULE>::validate_byte_slice(&bytes[#prev_offset_ident .. #prev_offset_ident + #size_ident])?;
-        };
-
-        prev_offset_ident = new_offset_ident;
-    }
-
-    (validators, prev_offset_ident)
+        quote!(<#ty as zerovec::ule::ULE>::validate_byte_slice(&bytes[#prev_offset_ident .. #prev_offset_ident + #size_ident])?;)
+    })
 }
 
 /// Make corresponding ULE fields for each field
